+++ scripts/deploy.sh
@@ -13,6 +13,7 @@ cleos wallet import --private-key 5KQwrPbwdL6PhXujxW37FSSQZ1JiwsST4cqQzDeyXtP79z
 echo -e "${CYAN}-----------------------CREATING ACCOUNTS-----------------------${NC}"
 cleos create account eosio eosio EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV
 cleos create account eosio eosio.wps EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV
+cleos create account eosio eosio.ramfee EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV
 cleos create account eosio eosio.token EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV
 cleos create account eosio myaccount EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV
 cleos create account eosio toaccount EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV
@@ -34,6 +35,7 @@ cleos push action eosio.token create '["eosio", "1000000000.0000 EOS"]' -p eosio
 cleos push action eosio.token issue '["eosio", "1000000000.0000 EOS", "init"]' -p eosio
 cleos transfer eosio myaccount "1000.0000 EOS"
 cleos transfer eosio toaccount "1000.0000 EOS"
+cleos transfer eosio eosio.ramfee "50000.0000 EOS"
 
 on_exit
 echo -e "${GREEN}--> Done${NC}"